Add patched aylet 0.5 for file conversion
The aylet player is patched so it writes my own .ay3 file format
from 3 channel AY files without playing them.

TODO: Make a nice converter out of the player without all
the unneeded stuff or provide an option for the player to have
both in one, or do it by symlink with argv[0] deciding which
function.

Signed-off-by: Andreas J. Reichel <andreas@reichel.bayern>